Changing name of new output group to PublishItemsOutputGroup and fixing tests
