+++ src/Tests/Microsoft.NET.TestFramework/CopyFilesTarget.cs
